# backEnd/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Social Media API",
        "description": "Documentation for Social Media API",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/api",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/titres": {
            "get": {
                "tags": [
                    "Titre"
                ],
                "summary": "Get all titres",
                "operationId": "fe884ebe14504f5ff365b906033039ac",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TitreResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Titre"
                ],
                "summary": "Create a new titre",
                "operationId": "8da2da0790927b2ce2eef240beb79381",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nomTitre": {
                                        "description": "Name of the titre",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description of the titre",
                                        "type": "string"
                                    },
                                    "number": {
                                        "description": "Number of the titre",
                                        "type": "integer"
                                    },
                                    "club_admin_id": {
                                        "description": "ID of the club admin",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Titre created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TitreResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/titres/{id}": {
            "get": {
                "tags": [
                    "Titre"
                ],
                "summary": "Get a titre by ID",
                "operationId": "93f1388b5f020c5f6375475903bc147a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the titre to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TitreResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Titre not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Titre"
                ],
                "summary": "Update a titre",
                "operationId": "c0e6b94d36e0f00ebb710e8ee244af6b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the titre to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nomTitre": {
                                        "description": "Name of the titre",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description of the titre",
                                        "type": "string"
                                    },
                                    "number": {
                                        "description": "Number of the titre",
                                        "type": "integer"
                                    },
                                    "club_admin_id": {
                                        "description": "ID of the club admin",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Titre updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TitreResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Titre not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Titre"
                ],
                "summary": "Delete a titre",
                "operationId": "550a14a5f36aa3a7a861f0a5ba761224",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the titre to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Titre deleted successfully"
                    },
                    "404": {
                        "description": "Titre not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/club-admins/{id}/titres": {
            "get": {
                "tags": [
                    "Titre"
                ],
                "summary": "Get titres by club admin ID",
                "operationId": "ec5e9622b796454cc2138cb4aeba99e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the club admin to retrieve titres for",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TitreResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Club admin not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TitreResource": {
                "title": "TitreResource",
                "description": "Resource representing a title won by a club admin",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nomTitre": {
                        "type": "string",
                        "example": "UEFA Champions League"
                    },
                    "description": {
                        "type": "string",
                        "example": "Won the UEFA Champions League in 2020"
                    },
                    "number": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}